+++ task/CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.0.3
 
-- Add tag editing UI
-- Add due date and priority to task list view
+- Add tag editing UI.
+- Add due date and priority to task list view.
 
 # 0.0.2